finish component re-alignment
