+++ src/form/Form.css
@@ -116,3 +116,7 @@ input:hover {
 .flower-four {
     padding: 11px 4px;
 }
+
+#name, #description {
+    padding-left: 5px;
+}
